+++ src/Yasumi/Provider/Germany/Berlin.php
@@ -12,8 +12,11 @@
 
 namespace Yasumi\Provider\Germany;
 
+use DateTime;
+use DateTimeZone;
 use Yasumi\Exception\InvalidDateException;
 use Yasumi\Exception\UnknownLocaleException;
+use Yasumi\Holiday;
 use Yasumi\Provider\Germany;
 
 /**
@@ -49,5 +52,43 @@ class Berlin extends Germany
         if ($this->year >= 2019) {
             $this->addHoliday($this->internationalWomensDay($this->year, $this->timezone, $this->locale));
         }
+
+        if ($this->year == 2020) {
+            $this->addHoliday($this->dayOfLiberation($this->timezone, $this->locale));
+        }
+    }
+
+    /**
+     * Day of Liberation
+     *
+     * Day of Liberation (Tag der Befreiung) is celebrated on May 8 2020 to commemorate the 75th anniversary
+     * of the German Instrument of Surrender.
+     *
+     * @link https://de.wikipedia.org/wiki/Tag_der_Befreiung
+     *
+     * @param string $timezone the timezone in which Day of Liberation is celebrated
+     * @param string $locale the locale for which Day of Liberation needs to be displayed in.
+     * @param string $type The type of holiday. Use the following constants: TYPE_OFFICIAL, TYPE_OBSERVANCE,
+     *                         TYPE_SEASON, TYPE_BANK or TYPE_OTHER. By default an official holiday is considered.
+     *
+     * @return Holiday
+     *
+     * @throws InvalidDateException
+     * @throws UnknownLocaleException
+     * @throws \InvalidArgumentException
+     * @throws \Exception
+     */
+    public function dayOfLiberation(
+        string $timezone,
+        string $locale,
+        string $type = Holiday::TYPE_OFFICIAL
+    ): Holiday {
+        return new Holiday(
+            'dayOfLiberation',
+            [],
+            new DateTime('2020-05-08', new DateTimeZone($timezone)),
+            $locale,
+            $type
+        );
     }
 }