+++ src/Yasumi/data/translations/dayOfLiberation.php
@@ -1,0 +1,17 @@
+<?php declare(strict_types=1);
+/**
+ * This file is part of the Yasumi package.
+ *
+ * Copyright (c) 2015 - 2019 AzuyaLabs
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Sacha Telgenhof <me@sachatelgenhof.com>
+ */
+
+// Translations for Day of Liberation
+return [
+    'de' => 'Tag der Befreiung',
+    'en' => 'Day of Liberation',
+];